+++ cvat/apps/engine/tests/test_rest_api.py
@@ -4811,6 +4811,11 @@ class TaskAnnotationAPITestCase(JobAnnotationAPITestCase):
                 annotations["shapes"] = points_wo_attrs \
                                       + rectangle_shapes_wo_attrs
 
+            elif annotation_format == "Open Images V6 1.0":
+                annotations["tags"] = tags_wo_attrs
+                annotations["shapes"] = rectangle_shapes_wo_attrs \
+                                      + polygon_shapes_wo_attrs
+
             elif annotation_format == "Market-1501 1.0":
                 tags_with_attrs = [{
                     "frame": 1,